// desktop/src/open-api/model/systemTaskStatus.ts
/**
 * Receipt Wrangler API.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type SystemTaskStatus = 'SUCCEEDED' | 'FAILED';

export const SystemTaskStatus = {
    Succeeded: 'SUCCEEDED' as SystemTaskStatus,
    Failed: 'FAILED' as SystemTaskStatus
};