// desktop/src/open-api/model/systemTaskType.ts
/**
 * Receipt Wrangler API.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type SystemTaskType = 'OCR_PROCESSING' | 'CHAT_COMPLETION' | 'MAGIC_FILL' | 'QUICK_SCAN' | 'EMAIL_READ' | 'EMAIL_UPLOAD' | 'SYSTEM_EMAIL_CONNECTIVITY_CHECK' | 'RECEIPT_PROCESSING_SETTINGS_CONNECTIVITY_CHECK' | 'RECEIPT_UPLOADED' | 'RECEIPT_UPDATED' | 'PROMPT_GENERATED' | 'API_KEY_DELETED';

export const SystemTaskType = {
    OcrProcessing: 'OCR_PROCESSING' as SystemTaskType,
    ChatCompletion: 'CHAT_COMPLETION' as SystemTaskType,
    MagicFill: 'MAGIC_FILL' as SystemTaskType,
    QuickScan: 'QUICK_SCAN' as SystemTaskType,
    EmailRead: 'EMAIL_READ' as SystemTaskType,
    EmailUpload: 'EMAIL_UPLOAD' as SystemTaskType,
    SystemEmailConnectivityCheck: 'SYSTEM_EMAIL_CONNECTIVITY_CHECK' as SystemTaskType,
    ReceiptProcessingSettingsConnectivityCheck: 'RECEIPT_PROCESSING_SETTINGS_CONNECTIVITY_CHECK' as SystemTaskType,
    ReceiptUploaded: 'RECEIPT_UPLOADED' as SystemTaskType,
    ReceiptUpdated: 'RECEIPT_UPDATED' as SystemTaskType,
    PromptGenerated: 'PROMPT_GENERATED' as SystemTaskType,
    ApiKeyDeleted: 'API_KEY_DELETED' as SystemTaskType
};